+++ code/Hyushik_TournMan_Web/Classes/ViewModels/AdminViewModels.cs
@@ -73,13 +73,13 @@ namespace Hyushik_TournMan_Web.Classes.ViewModels
     public class BreakingStoredValuesViewModel
     {
         [Range(minimum: 1, maximum: 20, ErrorMessage = "Count must be between 1 and 20.")]
-        public int BreakingMaxStationCount { get; set; }
+        public double BreakingMaxStationCount { get; set; }
         public string PossibleBoardWidths { get; set; }
         public string PossibleBoardDepths { get; set; }
 
         public double BreakingBoardExponent { get; set; }
-        public int BreakingMaximumBoards { get; set; }
-        public int BreakingMaximumAttempts { get; set; }
+        public double BreakingMaximumBoards { get; set; }
+        public double BreakingMaximumAttempts { get; set; }
         public double BreakingAttemptDecayRate { get; set; }
         public double BreakingSpacerPenalty { get; set; }
         public double BreakingPowerHoldPenalty { get; set; }
